+++ CHANGELOG.md
@@ -6,7 +6,7 @@ All Notable changes to `Csv` will be documented in this file
 
 ### Added
 
-- None
+- `EscapeFormula::unescapeRecord` does the opposite of `EscapeFormula::escapeRecord`
 
 ### Deprecated